+++ server/handler/update/update.go
@@ -130,6 +130,12 @@ func UpdateSpecificApp(c *gin.Context, repository db.AppRepository, db *mongo.Da
 	delete(ctxQueryMap, "id")
 
 	form, _ := c.MultipartForm()
+	checkAppVisibility, err := utils.CheckPrivate(ctxQueryMap["app_name"].(string), db, c)
+	if err != nil {
+		logrus.Error(err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to check private"})
+		return
+	}
 	var links []string
 	var extensions []string
 	var result bool
@@ -137,7 +143,7 @@ func UpdateSpecificApp(c *gin.Context, repository db.AppRepository, db *mongo.Da
 		files := form.File["file"] // Assuming the field name is "file" not "files"
 
 		for _, file := range files {
-			link, ext, err := utils.UploadToS3(ctxQueryMap, file, c, viper.GetViper())
+			link, ext, err := utils.UploadToS3(ctxQueryMap, file, c, viper.GetViper(), checkAppVisibility)
 			if err != nil {
 				logrus.Error(err)
 				c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to upload file to S3"})